Changes to reference functions and generating function in RCpp package.
In Rcpp branch, changed the reference functions to remove any allele
that has zero probability and to output a list of allele names to
correspond to the new lists of probabilities. This reduces workload
within the algorithm later. Changed mLoci and oLocus functions to handle
new references. Need to change Daisy still.