Massive Update to Initial Conditions
Since the switch in MGDrivE to a system similar to this, Sean solved the
initial critter distribution. That has been implemented in all 4 drives
available in mPlex, such that there is no burn-in now.
I do have concerns about equilibrium - it seems to equilibrate low, and
I'm not sure why. It may be nothing, or maybe a small bug causing a big
issue.