Reorganizing Shitty Code
There was a lot of repetitive code in the class initializers and reset
functions. Took all of it, made a few new functions, only call those
functions now. Only issue is that the family class uses a different set
of parameters, so had to keep 2 functions for the 2 different parameters
needed.