Several Updates and a New Problem
Tweaked all of the inheritance setups to do probabilities the same as
the current version of MGDrivE. The R part of appears to work fine.
Added mated female releases to the simulation. These have been tested
and work properly now.
There are extra alleles popping up. This occurs with mated female
relaeses and with male only releases. I'm not sure what's up yet, need
to figure that out before I move on.